factor OralHistoryContentL::OhmsXml::LegacyTranscript out of OralHistoryContentL::OhmsXml
In preparation for handling new style OHMS time-coded transcript too
